Share MARC logic with non-bib records
