+++ lib/sierra_postgres_utilities/version.rb
@@ -1,3 +1,3 @@
 module SierraPostgresUtilities
-  VERSION = '0.2.2'.freeze
+  VERSION = '0.2.3'.freeze
 end
